[NEW RBO] Add physical join builder
