+++ ydb/core/kqp/opt/rbo/kqp_rbo_physical_sort_builder.cpp
@@ -3,6 +3,65 @@ using namespace NYql::NNodes;
 using namespace NKikimr;
 using namespace NKikimr::NKqp;
 
+std::pair<TExprNode::TPtr, TVector<TExprNode::TPtr>> TPhysicalSortBuilder::BuildSortKeySelector(const TVector<TSortElement>& sortElements) {
+    auto arg = Build<TCoArgument>(Ctx, Pos).Name("arg").Done().Ptr();
+    TVector<TExprNode::TPtr> directions;
+    TVector<TExprNode::TPtr> members;
+
+    for (const auto& element : sortElements) {
+        // clang-format off
+        members.push_back(Build<TCoMember>(Ctx, Pos)
+            .Struct(arg)
+            .Name().Build(element.SortColumn.GetFullName())
+        .Done().Ptr());
+        // clang-format on
+
+        directions.push_back(Build<TCoBool>(Ctx, Pos).Literal().Build(element.Ascending ? "true" : "false").Done().Ptr());
+    }
+
+    TExprNode::TPtr selector;
+    if (sortElements.size() == 1) {
+        // clang-format off
+        selector = Build<TCoLambda>(Ctx, Pos)
+            .Args({arg})
+            .Body(members[0])
+            .Done().Ptr();
+        // clang-format on
+    } else {
+        // clang-format off
+        selector = Build<TCoLambda>(Ctx, Pos)
+            .Args({arg})
+            .Body<TExprList>().Add(members).Build()
+            .Done().Ptr();
+        // clang-format on
+    }
+
+    return std::make_pair(selector, directions);
+}
+
+TExprNode::TPtr TPhysicalSortBuilder::BuildSort(TExprNode::TPtr input, TOrderEnforcer& enforcer) {
+    if (enforcer.Action != EOrderEnforcerAction::REQUIRE) {
+        return input;
+    }
+
+    auto [selector, dirs] = BuildSortKeySelector(enforcer.SortElements);
+
+    TExprNode::TPtr dirList;
+    if (dirs.size() == 1) {
+        dirList = dirs[0];
+    } else {
+        dirList = Build<TExprList>(Ctx, Pos).Add(dirs).Done().Ptr();
+    }
+
+    // clang-format off
+    return Build<TCoSort>(Ctx, Pos)
+        .Input(input)
+        .SortDirections(dirList)
+        .KeySelectorLambda(selector)
+    .Done().Ptr();
+    // clang-format on
+}
+
 TVector<TExprNode::TPtr> TPhysicalSortBuilder::BuildSortKeysForWideSort(const TVector<TInfoUnit>& inputs, const TVector<TSortElement>& sortElements) {
     // We have to map wide input with sort elements to find a right index.
     THashMap<TString, ui32> indices;
