+++ ydb/core/kqp/opt/rbo/ya.make
@@ -17,6 +17,7 @@ SRCS(
     kqp_prune_columns_stage.cpp
     kqp_rbo_physical_aggregation_builder.cpp
     kqp_rbo_physical_sort_builder.cpp
+    kqp_rbo_physical_join_builder.cpp
     kqp_rbo_physical_convertion_utils.cpp
 )
